refactor(http): extract plane-independent server core
