+++ lib/llm/src/discovery/watcher.rs
@@ -10,6 +10,7 @@ use tokio::task::JoinHandle;
 
 use anyhow::Context as _;
 use dashmap::{DashMap, DashSet};
+use dynamo_http_server::metrics::Metrics;
 use dynamo_kv_router::PrefillLoadEstimator;
 use futures::StreamExt;
 
@@ -32,7 +33,6 @@ use crate::{
     backend::Backend,
     discovery::{KvWorkerMonitor, WORKER_TYPE_DECODE, WorkerSet},
     entrypoint::{self, ChatEngineFactoryCallback, RouterConfig},
-    http::service::metrics::Metrics,
     kv_router::PrefillRouter,
     local_model::runtime_config::TokenizerBackend,
     model_card::ModelDeploymentCard,
