+++ lib/llm/src/entrypoint/input/http.rs
@@ -181,7 +181,7 @@ async fn run_watcher(
     migration_max_seq_len: Option<u32>,
     namespace_filter: NamespaceFilter,
     http_service: Arc<HttpService>,
-    metrics: Arc<crate::http::service::metrics::Metrics>,
+    metrics: Arc<dynamo_http_server::metrics::Metrics>,
     chat_engine_factory: Option<ChatEngineFactoryCallback>,
     prefill_load_estimator: Option<Arc<dyn dynamo_kv_router::PrefillLoadEstimator>>,
     local_model_path: Option<PathBuf>,
@@ -268,12 +268,14 @@ fn update_http_endpoints(service: Arc<HttpService>, model_type: ModelUpdate) {
 /// Updates metrics for model type changes
 fn update_model_metrics(
     model_type: ModelUpdate,
-    metrics: Arc<crate::http::service::metrics::Metrics>,
+    metrics: Arc<dynamo_http_server::metrics::Metrics>,
 ) {
     match model_type {
         ModelUpdate::Added(card) => {
             tracing::debug!("Updating metrics for added model: {}", card.display_name);
-            if let Err(err) = metrics.update_metrics_from_mdc(&card) {
+            if let Err(err) =
+                crate::http::service::metrics::update_metrics_from_mdc(&metrics, &card)
+            {
                 tracing::warn!(%err, model_name=card.display_name, "update_metrics_from_mdc failed");
             }
         }
