+++ lib/llm/src/frontend_config.rs
@@ -17,7 +17,7 @@ use dynamo_runtime::config::{
 ///
 /// Contains the optional metric name prefix resolved from `--metrics-prefix` or
 /// `DYN_METRICS_PREFIX`. HTTP services use it when constructing
-/// `http::service::metrics::Metrics`; gRPC mode also exposes the prefix through
+/// `dynamo_http_server::metrics::Metrics`; gRPC mode also exposes the prefix through
 /// the existing `LocalModel::metrics_prefix()` compatibility accessor.
 #[derive(Debug, Clone, PartialEq, Eq)]
 pub struct MetricsConfig {
