+++ lib/llm/src/grpc/service/openai.rs
@@ -1,6 +1,7 @@
 // SPDX-FileCopyrightText: Copyright (c) 2024-2026 NVIDIA CORPORATION & AFFILIATES. All rights reserved.
 // SPDX-License-Identifier: Apache-2.0
 
+use dynamo_http_server::disconnect::{ConnectionHandle, create_connection_monitor};
 use dynamo_runtime::{
     engine::AsyncEngineContext,
     pipeline::{AsyncEngineContextProvider, Context},
@@ -20,10 +21,8 @@ use super::kserve;
 use super::kserve::inference;
 
 // [gluo NOTE] These are common utilities that should be shared between frontends
-use crate::http::service::{
-    disconnect::{ConnectionHandle, create_connection_monitor},
-    metrics::{CancellationLabels, Endpoint, InflightGuard, process_response_and_observe_metrics},
-};
+use crate::http::service::metrics::process_response_and_observe_metrics;
+use dynamo_http_server::metrics::{CancellationLabels, Endpoint, InflightGuard};
 use dynamo_protocols::types::{CompletionFinishReason, CreateCompletionRequest, Prompt};
 
 use tonic::{Status, metadata::MetadataMap};
@@ -112,10 +111,10 @@ pub async fn completion_response_stream(
 
     // issue the generate call on the engine
     let stream = engine.generate(request).await.map_err(|e| {
-        if crate::http::service::metrics::request_was_rejected(e.as_ref()) {
+        if dynamo_http_server::metrics::request_was_rejected(e.as_ref()) {
             state.metrics_clone().inc_rejection(
                 &model_name,
-                crate::http::service::metrics::Endpoint::Completions,
+                dynamo_http_server::metrics::Endpoint::Completions,
             );
             return Status::resource_exhausted(e.to_string());
         }
