+++ lib/llm/src/protocols/common/timing.rs
@@ -16,17 +16,16 @@ use serde::{Deserialize, Serialize};
 use tokio::sync::{OwnedSemaphorePermit, Semaphore};
 use utoipa::ToSchema;
 
-use crate::http::service::metrics::{
+use crate::protocols::common::extensions::WorkerIdInfo;
+use dynamo_http_server::metrics::{
     WORKER_LAST_INPUT_SEQUENCE_TOKENS_GAUGE, WORKER_LAST_INTER_TOKEN_LATENCY_GAUGE,
     WORKER_LAST_TIME_TO_FIRST_TOKEN_GAUGE,
 };
-use crate::protocols::common::extensions::WorkerIdInfo;
 
 /// Worker type constants for Prometheus metric labels.
 /// These are stored in RequestTracker at routing time to avoid costly MDC lookups
 /// when updating per-worker metrics (TTFT, ITL).
-pub const WORKER_TYPE_PREFILL: &str = "prefill";
-pub const WORKER_TYPE_DECODE: &str = "decode";
+pub use dynamo_http_server::metrics::{WORKER_TYPE_DECODE, WORKER_TYPE_PREFILL};
 const UNSET_DP_RANK_LABEL: &str = "none";
 
 /// Phase of the request in disaggregated serving.
