+++ lib/llm/tests/http_metrics.rs
@@ -3,9 +3,10 @@
 
 use anyhow::Error;
 use async_stream::stream;
+use dynamo_http_server::metrics::Endpoint;
 use dynamo_llm::{
     discovery::UNKNOWN_METRIC_MODEL,
-    http::service::{metrics::Endpoint, service_v2::HttpService},
+    http::service::service_v2::HttpService,
     model_card::ModelDeploymentCard,
     preprocessor::LLMMetricAnnotation,
     protocols::{
@@ -539,11 +540,10 @@ mod integration_tests {
         // This simulates what the ModelWatcher polling task would do in production
         let _ = manager.save_model_card("test-mdc-key", card.clone());
 
-        if let Err(e) = service
-            .state()
-            .metrics_clone()
-            .update_metrics_from_mdc(&card)
-        {
+        if let Err(e) = dynamo_llm::http::service::metrics::update_metrics_from_mdc(
+            &service.state().metrics_clone(),
+            &card,
+        ) {
             tracing::debug!(
                 model = %card.display_name,
                 error = %e,
